+++ src/Capture_the_Ether/Lotteries/Guess_the_new_number/GuessTheNewNumberChallenge.sol
@@ -13,7 +13,7 @@ contract GuessTheNewNumberChallenge {
 
     function guess(uint8 n) public payable {
         require(msg.value == 1 ether);
-        uint8 answer = uint8(uint(keccak256(abi.encodePacked(blockhash(block.number - 1), block.timestamp))));
+        uint8 answer = uint8(uint256(keccak256(abi.encodePacked(blockhash(block.number - 1), block.timestamp))));
 
         if (n == answer) {
             payable(msg.sender).transfer(2 ether);
